LayoutTests/imported/w3c/web-platform-tests/webcodecs/audio-encoder-codec-specific.https.any-expected.txt


PASS Test the Opus DTX flag works.
FAIL Test the Opus bitrateMode flag works. assert_less_than: expected a number less than 40330 but got 80660
PASS Test the AAC bitrateMode flag works.
